FLIZ-151 rename: regsiter 관련 컴포넌트 위치 이동
